rename File into NexusFile
