+++ nexus-sdk-js/src/index.ts
@@ -6,7 +6,7 @@ export { default as Project } from './Project';
 export { default as Resource } from './Resource';
 export { default as ACL } from './ACL';
 export { default as Realm } from './Realm';
-export { default as File } from './File';
+export { default as NexusFile } from './File';
 export * from './View';
 export * from './utils/types';
